Rename to ApartmentFilter
